voicera_backend/requirements.txt
pymongo>=4.6.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
mailtrap>=2.0.0
minio>=7.2.0
httpx>=0.25.0

# RAG / knowledge (rag_system scripts + rag_server)
chromadb>=0.4.0
openai>=1.0.0
numpy>=1.24.0
pypdf>=4.0.0

